+++ src/repeatedTask.cpp
@@ -159,14 +159,14 @@ void PhrasedRepeatedTask::addRepeatedTaskToSEDML(SedDocument* sedml) const
   sedRepeatedTask->setResetModel(m_resetModel);
   for (size_t t=0; t<m_tasks.size(); t++) {
     SedSubTask* subtask = sedRepeatedTask->createSubTask();
-    subtask->setOrder(t);
+    subtask->setOrder((int)t);
     subtask->setTask(m_tasks[t]);
   }
   for (size_t c=0; c<m_changes.size(); c++) {
     m_changes[c].addModelChangeToSEDMLRepeatedTask(sedRepeatedTask, m_tasks);
   }
   //Now that all the changes have been added, we make sure that all the math variables can be found
-  for (size_t sv=0; sv<sedRepeatedTask->getNumTaskChanges(); sv++) {
+  for (unsigned int sv=0; sv<sedRepeatedTask->getNumTaskChanges(); sv++) {
     SedSetValue* ssv = sedRepeatedTask->getTaskChange(sv);
     addLocalVariablesToSetValue(ssv, sedRepeatedTask);
   }
